requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via kopf
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
attrs==25.3.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   kopf
    #   uvicorn
dnspython==2.7.0
    # via pymongo
durationpy==0.10
    # via kubernetes
fastapi==0.115.13
    # via shield-backend (pyproject.toml)
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.40.3
    # via kubernetes
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
iso8601==2.1.0
    # via kopf
kopf==1.38.0
    # via shield-backend (pyproject.toml)
kubernetes==33.1.0
    # via shield-backend (pyproject.toml)
multidict==6.5.1
    # via
    #   aiohttp
    #   yarl
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   shield-backend (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
pymongo==4.13.2
    # via shield-backend (pyproject.toml)
python-dateutil==2.9.0.post0
    # via kubernetes
python-dotenv==1.1.1
    # via shield-backend (pyproject.toml)
python-json-logger==3.3.0
    # via kopf
pyyaml==6.0.2
    # via
    #   kopf
    #   kubernetes
requests==2.32.4
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   kubernetes
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.14.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
uvicorn==0.34.3
    # via shield-backend (pyproject.toml)
websocket-client==1.8.0
    # via kubernetes
yarl==1.20.1
    # via aiohttp
